Initial go-round of uoconvert smartness - `SystemState::tiles` changed from c-style `Tile*` to `std::vector<>` - `MaxTileId` removed from pol.cfg, uoconvert command line - Calculate `MaxTileId` from...   - (pol) maximum defined graphic in tiles.cfg   - (uoconvert) tiledata.mul byte size - `UseNewHSAFormat` removed from uoconvert.cfg - Calculated `UseNewHSAFormat` from...   - (pol) not needed   - (uoconvert) tiledata.mul byte size - Auto-calculate dimenions for maps 0,1 if not provided from byte size

+++ pol-core/pol/polcfg.h
@@ -59,7 +59,7 @@ struct PolConfig
   unsigned short inactivity_warning_timeout;
   unsigned short inactivity_disconnect_timeout;
   unsigned short min_cmdlevel_to_login;
-  unsigned int max_tile_id;
+  size_t max_tile_id;
   unsigned int max_objtype;
 
   unsigned short max_clients;
